requirements.txt
click>=8.1
paramiko>=3.4
pyyaml>=6.0
requests>=2.31
rich>=13.7
cryptography>=42.0
mcp>=0.1.0
